requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
PyPDF2==3.0.1

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# OCR
easyocr==1.7.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0